Create separate `phylum_project` crate
To ensure all Phylum projects using the `.phylum_project` file have a
common understanding of its content, the parsing for this file has been
extracted into a separate crate.

Closes #908.
