+++ cli/Cargo.toml
@@ -36,6 +36,7 @@ log = "0.4.6"
 maplit = "1.0.2"
 open = "3.0.1"
 phylum_lockfile = { path = "../lockfile" }
+phylum_project = { path = "../phylum_project" }
 phylum_types = { git = "https://github.com/phylum-dev/phylum-types", branch = "development" }
 prettytable-rs = "0.10.0"
 rand = "0.8.4"
