+++ phylum_project/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "phylum_project"
+description = "Phylum project configuration handling"
+repository = "https://github.com/phylum-dev/cli"
+version = "0.1.0"
+authors = ["Phylum, Inc. <engineering@phylum.io>"]
+license = "GPL-3.0-or-later"
+edition = "2021"
+rust-version = "1.64.0"
+
+[dependencies]
+phylum_types = { git = "https://github.com/phylum-dev/phylum-types", branch = "development" }
+chrono = { version = "^0.4", default-features = false, features = ["serde", "clock"] }
+serde = { version = "1.0.144", features = ["derive"] }
+serde_yaml = "0.9.2"
+log = "0.4.6"
